Adding licensing information for Keras models in code (#83)
Adding licensing information for Keras models in code

This PR adds specific license information for each model used in DeepImageFeaturizer. It essentially says that the models come via Keras, states the Keras license information, and for each model that comes from another source adds the licensing information for the source.