Fix bug in conversions from row image to/from BufferedImage (#99)
Fix bug in conversions to/from BufferedImage in which we copied raw byte data to BufferedImage rasters using the wrong channel ordering. The fix in this PR is to use BufferedImage.setRGB, BufferedImage.getRGB APIs instead of accessing image raster data directly for three and four-channel images.

Also enhanced an existing unit test to verify that we correctly convert from row image to BufferedImage for one and four-channel images.